requirements.txt
torch>=2.0
torchvision
numpy
Pillow
open3d
huggingface_hub
einops
safetensors
pytorch_metric_learning
pytorch-lightning
termcolor
viser==0.2.23
tqdm
omegaconf
opencv-python
scipy
requests
trimesh
matplotlib
lz4
